using group.children instead of the query selector
Unfortunately, there isn't a querySelectorAll that works for getting the immediate children of an element. :scope would work but it has limited support and I dont want to add a polyfill. Using .children and then filtering out the things that don't have a pfe-heading attribute should do the trick.

Fixes #314