moved the creation of PfeNavigation to top of the list
I'm not really sure why this fixes things but order seems to matter here. I found the issue by getting rid of PfeNavigationItem and noticed that when I brought it back in, it created an error with the polyfill. Moving PfeNavigation ahead of it in the creation process seems to fix the issue.