fix: on mutation, cascade to new elements only
When a cascade is triggered by a mutation observer, copy attributes down
to the newly created elements only.  Previously, I had code implemented
that tried to do this but it wasn't working.  It was finding which
cascade selectors matched the new elements, but then copied the
attributes to _all_ elements that matched those selectors.