fix: on mutation, apply cascade to new elements only (#1440)
* fix: on mutation, cascade to new elements only

When a cascade is triggered by a mutation observer, copy attributes down
to the newly created elements only.  Previously, I had code implemented
that tried to do this but it wasn't working.  It was finding which
cascade selectors matched the new elements, but then copied the
attributes to _all_ elements that matched those selectors.

* update changelog

* move IE11 observer reattach to same branch as detach

* update changelog

* Update CHANGELOG-1.x.md

* fix incorrect iteration over cascading attrNames

* Update CHANGELOG-1.x.md

Co-authored-by: [ Cassondra ] <castastrophe@users.noreply.github.com>

* add jsdoc params to _cascadeAttribute

* Update elements/pfelement/src/pfelement.js

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: [ Cassondra ] <castastrophe@users.noreply.github.com>