Removing checks for copying text content of HTML
Discussed with Kyle Buchanan, Michael Clayton, and other CPFED folks, guarding against the HTML element that will have it's text copied isn't really worth it. There isn't a good way to know that it _should_ be disabled and then _should_ be enabled later.
Better to leave it to the click event, and hopefully things were setup so it'll work when the user clicks on it