Simplifying back button text and removing unused functions
Back button text did not need extra markup, other than the surrrounding button. Moved it to a internal property.
Removed unused functions made during original development, but are now unused
Updated Demo HTML, thought text in 'details' was a bug in functionality, but was content in the slot. Hopefully the slight adjustment makes that more clear.